refactor(#138): extract export.py (module-split phase, ADR 0005)
First module-split step: move the SVG/DXF/PDF export and post-processing free
functions out of make_drawing.py into a dedicated export.py — fix_svg_page_size,
add_svg_hyperlink, add_svg_metadata, set_dxf_metadata, sanitize_svg_arcs,
_render_pdf, _export_shape (+ _elements and the arc-regex/constants). Drawing's
export() / export_pdf() / _add_shapes() stay as the public wrappers that drive
them. make_drawing.py drops ~270 lines (3762, from ~4030 after Step 3).

Pure code move — no behaviour change:
- export.py mostly consumes drawing contents (paths, exporters, shapes); it
  depends only on _core (the shared _DRAFTWRIGHT_URL) and build123d, so it sits
  below make_drawing in the DAG (no cycle).
- make_drawing re-imports the symbols its Drawing methods use, so existing
  `from draftwright.make_drawing import …` (and the public fix_svg_page_size via
  __init__) keep working. Two test imports were deliberately redirected to the
  new home (draftwright.export) — the _MIN_ARC_RADIUS/_SVG_ARC_RE and the
  _DRAFTWRIGHT_URL/add_svg_hyperlink imports.
- The golden gate passes against the committed snapshots with no regeneration —
  the SVG/DXF digests are byte-identical — and the export-path tests (hyperlink,
  metadata, arc sanitisation, shape-export fallback, PDF link, e2e SVG standards)
  pass.

CLAUDE.md module list + DAG updated.
