+++ src/draftwright/export.py
@@ -1,0 +1,257 @@
+"""SVG / DXF / PDF export and post-processing (#138 / ADR 0005, Step "export").
+
+The free functions the public `Drawing.export()` / `Drawing.export_pdf()` wrappers
+drive: SVG page-size fixing, the attribution hyperlink and metadata, DXF metadata,
+near-degenerate arc sanitisation, the element-wise shape-export degradation, and
+the cairo PDF render. Moved out of `make_drawing.py` so the orchestration there
+stays thin; this module mostly *consumes* drawing contents (paths, exporters,
+shapes), so it sits below `make_drawing` in the import DAG and depends only on
+`_core` (the shared `_DRAFTWRIGHT_URL`) and build123d.
+"""
+
+from __future__ import annotations
+
+import logging
+import re
+from pathlib import Path
+
+from build123d import ExportSVG
+
+from draftwright._core import _DRAFTWRIGHT_URL
+
+_log = logging.getLogger(__name__)
+
+
+def fix_svg_page_size(svg_path: str, page_w: float, page_h: float) -> None:
+    """Rewrite the SVG width/height/viewBox to match the full ISO page size.
+
+    ExportSVG crops to content bounding box; this expands it to the declared
+    page so the rendering fills the correct A-series sheet.
+    """
+    data = Path(svg_path).read_text(encoding="utf-8")
+    data = re.sub(r'width="[^"]*"', f'width="{page_w:.3f}mm"', data, count=1)
+    data = re.sub(r'height="[^"]*"', f'height="{page_h:.3f}mm"', data, count=1)
+    data = re.sub(
+        r'viewBox="[^"]*"',
+        f'viewBox="0 -{page_h:.3f} {page_w:.3f} {page_h:.3f}"',
+        data,
+        count=1,
+    )
+    Path(svg_path).write_text(data, encoding="utf-8")
+
+
+# Below this, an elliptical-arc radius (page-mm) is treated as degenerate.
+# Real feature arcs are orders of magnitude larger; the bad ones are ~1e-7.
+_MIN_ARC_RADIUS = 1e-3
+
+_SVG_NUM = r"(-?\d+\.?\d*(?:[eE][-+]?\d+)?)"
+_SVG_ARC_RE = re.compile(
+    r"A\s*"
+    + _SVG_NUM
+    + r"[ ,]+"
+    + _SVG_NUM
+    + r"[ ,]+"
+    + _SVG_NUM
+    + r"[ ,]+([01])[ ,]*([01])[ ,]+"
+    + _SVG_NUM
+    + r"[ ,]+"
+    + _SVG_NUM
+)
+
+
+def add_svg_hyperlink(svg_path: str, rect, href: str = _DRAFTWRIGHT_URL) -> None:
+    """Overlay a transparent, clickable hyperlink rectangle on the SVG.
+
+    *rect* is ``(x0, y0, x1, y1)`` in drawing page coordinates (Y up). The SVG
+    viewBox is ``0 -page_h page_w page_h`` (see :func:`fix_svg_page_size`), so a
+    page point ``(X, Y)`` maps to SVG ``(X, -Y)``. The rect is wrapped in an
+    ``<a>`` with a fully transparent fill and ``pointer-events="all"`` so the
+    whole cell is clickable without altering the rendered drawing. DXF carries
+    no equivalent, so this is SVG-only.
+    """
+    x0, y0, x1, y1 = rect
+    sx, sy, sw, sh = x0, -y1, x1 - x0, y1 - y0
+    link = (
+        f'<a xlink:href="{href}" href="{href}" target="_blank">'
+        f'<rect x="{sx:.3f}" y="{sy:.3f}" width="{sw:.3f}" height="{sh:.3f}" '
+        f'fill="transparent" pointer-events="all"/></a>'
+    )
+    data = Path(svg_path).read_text(encoding="utf-8")
+    if f'href="{href}"' in data:
+        return  # already injected — keep idempotent like add_svg_metadata
+    if "xmlns:xlink" not in data:
+        data = re.sub(r"(<svg\b)", r'\1 xmlns:xlink="http://www.w3.org/1999/xlink"', data, count=1)
+    data = data.replace("</svg>", link + "</svg>", 1)
+    Path(svg_path).write_text(data, encoding="utf-8")
+
+
+_GENERATED_BY = "Generated by draftwright"
+
+
+def add_svg_metadata(svg_path: str) -> None:
+    """Embed a 'generated by draftwright' note in the SVG (comment + Dublin Core
+    ``<metadata>``), so the attribution survives in the file itself."""
+    meta = (
+        f"<!-- {_GENERATED_BY}: {_DRAFTWRIGHT_URL} -->"
+        "<metadata>"
+        '<rdf:RDF xmlns:rdf="http://www.w3.org/1999/02/22-rdf-syntax-ns#" '
+        'xmlns:dc="http://purl.org/dc/elements/1.1/">'
+        "<rdf:Description>"
+        "<dc:creator>draftwright</dc:creator>"
+        f"<dc:source>{_DRAFTWRIGHT_URL}</dc:source>"
+        f"<dc:description>{_GENERATED_BY}</dc:description>"
+        "</rdf:Description></rdf:RDF></metadata>"
+    )
+    data = Path(svg_path).read_text(encoding="utf-8")
+    if "<dc:creator>draftwright</dc:creator>" not in data:
+        data = re.sub(r"(<svg\b[^>]*>)", r"\1" + meta, data, count=1)
+        Path(svg_path).write_text(data, encoding="utf-8")
+
+
+def set_dxf_metadata(dxf) -> None:
+    """Stamp 'generated by draftwright' into the DXF header as custom properties
+    (``$CUSTOMPROPERTY``), which CAD applications surface in drawing properties.
+    Best-effort: a no-op if the exporter doesn't expose an ezdxf document."""
+    doc = getattr(dxf, "_document", None)
+    if doc is None:
+        return
+    try:
+        cv = doc.header.custom_vars
+        cv.append("GeneratedBy", "draftwright")
+        cv.append("draftwrightURL", _DRAFTWRIGHT_URL)
+    except Exception:
+        pass
+
+
+def sanitize_svg_arcs(svg_path: str) -> int:
+    """Rewrite near-degenerate elliptical arcs as straight line segments.
+
+    build123d's ``ExportSVG`` projects a circular edge seen edge-on (a hole or
+    fillet rim whose plane is parallel to the view direction) as an elliptical
+    arc with a vanishing minor radius (``ry`` ≈ 1e-7).  The SVG spec says a
+    zero-radius arc is a straight line, but because the radius is not *exactly*
+    zero, renderers (librsvg, cairosvg) treat it as a hugely eccentric ellipse
+    and draw a spurious full-page line.  Each such arc (``A rx ry rot lf sf x
+    y``) with ``rx`` or ``ry`` below :data:`_MIN_ARC_RADIUS` is replaced by
+    ``L x y`` — its true geometry.  Returns the number of arcs rewritten.
+    """
+    data = Path(svg_path).read_text(encoding="utf-8")
+    n = 0
+
+    def _repl(m):
+        nonlocal n
+        if abs(float(m.group(1))) < _MIN_ARC_RADIUS or abs(float(m.group(2))) < _MIN_ARC_RADIUS:
+            n += 1
+            return f"L {m.group(6)} {m.group(7)}"
+        return m.group(0)
+
+    fixed = _SVG_ARC_RE.sub(_repl, data)
+    if n:
+        Path(svg_path).write_text(fixed, encoding="utf-8")
+    return n
+
+
+def _render_pdf(svg_path: str, pdf_path: str, page_h: float, link_rect=None) -> None:
+    """Render *svg_path* to *pdf_path* via cairosvg, adding draftwright metadata
+    and — when *link_rect* (drawing page coords, Y up) is given — a clickable
+    PDF link annotation over that rectangle.
+
+    cairosvg does not translate the SVG ``<a>`` element into a PDF link, so the
+    link is added with cairo's ``CAIRO_TAG_LINK``. The drawing page maps to PDF
+    points at 72/25.4 pt·mm⁻¹; PDF user space is top-left origin (Y down) while
+    the page is Y up, so a page rect ``(x0, y0, x1, y1)`` has its top edge at
+    ``page_h - y1``. Falls back to a plain render if the richer path is
+    unavailable for any reason."""
+    import cairosvg
+
+    try:
+        import cairocffi
+        from cairosvg.parser import Tree
+        from cairosvg.surface import PDFSurface
+    except Exception:
+        cairosvg.svg2pdf(url=svg_path, write_to=pdf_path)
+        return
+
+    try:
+        instance = PDFSurface(Tree(url=svg_path), pdf_path, dpi=96)
+        surface = instance.cairo
+        try:
+            surface.set_metadata(cairocffi.PDF_METADATA_CREATOR, _GENERATED_BY)
+            surface.set_metadata(cairocffi.PDF_METADATA_TITLE, _GENERATED_BY)
+        except Exception:
+            pass  # older cairo without per-field metadata
+        if link_rect is not None:
+            k = 72.0 / 25.4
+            x0, y0, x1, y1 = link_rect
+            rx, ry, rw, rh = x0 * k, (page_h - y1) * k, (x1 - x0) * k, (y1 - y0) * k
+            ctx = cairocffi.Context(surface)
+            ctx.identity_matrix()
+            ctx.tag_begin(
+                cairocffi.TAG_LINK,
+                f"uri='{_DRAFTWRIGHT_URL}' rect=[{rx:.2f} {ry:.2f} {rw:.2f} {rh:.2f}]",
+            )
+            ctx.tag_end(cairocffi.TAG_LINK)
+        instance.finish()
+    except Exception:
+        # Never fail the export over the link/metadata extras; degrade to a
+        # plain render. Logged at debug so a regression in the link annotation
+        # is diagnosable (in normal use only the dedicated test would catch it).
+        _log.debug("PDF link/metadata extras failed; rendered a plain PDF", exc_info=True)
+        cairosvg.svg2pdf(url=svg_path, write_to=pdf_path)
+
+
+def _elements(shape):
+    """Decompose *shape* for export retry: faces plus any loose edges."""
+    faces = list(shape.faces())
+    if not faces:
+        return list(shape.edges())
+    owned = {e for f in faces for e in f.edges()}
+    return faces + [e for e in shape.edges() if e not in owned]
+
+
+def _export_shape(exporter, shape, layer, ctx):
+    """Add *shape* to *exporter*, degrading element-by-element on failure.
+
+    build123d's exporters abort the whole export on the first edge whose
+    curve cannot be approximated (a bare ``AssertionError`` from OCCT, #83).
+    Instead, drop only the offending elements with a warning naming the
+    view/layer, and raise (with that context) only if nothing exported.
+
+    ``ExportSVG.add_shape`` is atomic — it appends converted elements only
+    after the whole shape succeeds — so the shape is tried in one call first.
+    ``ExportDXF`` writes edge-by-edge as it converts, so a mid-shape failure
+    would leave partial output that a blind retry duplicates; for it (and any
+    unknown exporter) every element is added individually from the start.
+    """
+    first_err = None
+    if isinstance(exporter, ExportSVG):
+        try:
+            exporter.add_shape(shape, layer=layer)
+            return
+        except Exception as exc:
+            first_err = exc
+            _log.warning(
+                "%s (layer %r) failed to export as one shape: %s — retrying element-wise",
+                ctx,
+                layer,
+                exc,
+            )
+    elements = _elements(shape)
+    skipped = 0
+    for element in elements:
+        try:
+            exporter.add_shape(element, layer=layer)
+        except Exception as exc:
+            first_err = first_err or exc
+            skipped += 1
+            _log.debug("%s (layer %r): element failed to convert: %s", ctx, layer, exc)
+    if skipped == len(elements) and first_err is not None:
+        raise RuntimeError(f"{ctx} (layer {layer!r}): nothing could be exported") from first_err
+    if skipped:
+        _log.warning(
+            "%s (layer %r): skipped %d of %d elements that failed to convert",
+            ctx,
+            layer,
+            skipped,
+            len(elements),
+        )
