+++ tests/test_e2e_standards.py
@@ -107,7 +107,7 @@ _CTC_AP242_OK = ["01", "02", "03", "04", "05"]
 @pytest.mark.timeout(600)
 @pytest.mark.parametrize("n", _CTC_AP203_OK)
 def test_ctc_ap203_meets_standards_no_degenerate_arcs(tmp_path, n):
-    from draftwright.make_drawing import _MIN_ARC_RADIUS, _SVG_ARC_RE
+    from draftwright.export import _MIN_ARC_RADIUS, _SVG_ARC_RE
 
     step = FIXTURES / f"nist_ctc_{n}_asme1_ap203.stp"
     stem = str(tmp_path / f"ctc{n}")
